removed some old versions of IsValidCoord and IsValidLineSegment, added a new heuristic. I need to test it and compare it to the old heuristic
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@22649 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 288e2228e0e6b3ab8e6f5ce85041aee888bb9281

+++ motion_planning/planning_research/sbpl_arm_planner/include/sbpl_arm_planner/robarm3d/environment_robarm3d.h
@@ -258,7 +258,7 @@ typedef struct ENV_ROBARM_CONFIG
     double grasped_object_length_m;
     bool use_voxel3d_occupancy_grid;
 		double RPYHeuristicDistance_m;
-		double RPYHeuristicDistance_c;
+		int RPYHeuristicDistance_c;
     bool use_selective_actions;
     bool exact_gripper_collision_checking;
     bool use_jacobian_motion_prim;
